requirements.txt
Pillow==10.1.0
PyYAML==6.0.1
numpy==1.25.2
fonttools==4.43.1
click==8.1.7